essentially back to where i was in 32 bit mode. now lets map the kernel.

+++ kernel/interrupts/irq_handlers/handlers.h
@@ -32,9 +32,9 @@ namespace irq_handlers
     };
     struct __attribute__((packed)) interrupt_frame_t
     {
-        uint32_t eip;
-        uint32_t cs;
-        uint32_t flags;
+        uint64_t eip;
+        uint64_t cs;
+        uint64_t flags;
     };
     void handle_divide_by_zero(interrupt_frame_t *frame);
     void handle_timer(interrupt_frame_t *frame);
